requirements-dev.txt
black==21.9b0
django-debug-toolbar
mock
pycodestyle==2.7.0
pyflakes==2.3.1
